+++ src/frontend/apps/impress/src/features/auth/__tests__/utils.test.tsx
@@ -4,11 +4,6 @@ import { afterEach, describe, expect, it, vi } from 'vitest';
 import { SILENT_LOGIN_RETRY } from '../conf';
 import { gotoLogout, gotoSilentLogin } from '../utils';
 
-// Mock the Crisp service
-vi.mock('@/services/Crisp', () => ({
-  terminateCrispSession: vi.fn(),
-}));
-
 // Add mock on window.location.replace
 const mockReplace = vi.fn();
 Object.defineProperty(window, 'location', {
@@ -32,12 +27,9 @@ describe('utils', () => {
     localStorage.clear();
   });
 
-  it('checks support session is terminated when logout', async () => {
-    const { terminateCrispSession } = await import('@/services/Crisp');
-
+  it('checks logout redirects to logout URL', async () => {
     gotoLogout();
 
-    expect(terminateCrispSession).toHaveBeenCalled();
     expect(mockReplace).toHaveBeenCalledWith(
       'http://test.jest/api/v1.0/logout/',
     );