+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to
 - ♿️(frontend) hide mobile left panel from screen readers when collapsed #2450
 - ♿️(frontend) enable blocknote heading ids for toc anchors #2449
 - ♿️(frontend) focus export modal on format select #2421
+- ♿️(frontend) configurable legal submenu in HelpMenu, remove Crisp #2416
 
 ## [v5.3.0] - 2026-06-19